don't draw text at fractional pixel positions (at least at base resolution)
git-svn-id: https://pioneer.svn.sourceforge.net/svnroot/pioneer/trunk@325 e632f14b-6550-0410-b89e-a82653faca30

+++ src/GuiVScrollPortal.cpp
@@ -84,7 +84,7 @@ void VScrollPortal::Draw()
 	if (toScroll < 0) toScroll = 0;
 
 	glPushMatrix();
-	glTranslatef(0, -m_scrollY*toScroll, 0);
+	glTranslatef(0, (int)(-m_scrollY*toScroll), 0);
 	Container::Draw();
 	glPopMatrix();
 	EndClipping();
